Fix crappy memory management.
We can’t really be passing ishake_block objects around like that if we are using threads, specially because they contain pointers to the data. Additionally, given that we don’t know when a block has been processed, we need to move the responsibility to free unused memory to the worker threads themselves. This means we cannot push a block to the processing queue, change something, and push it again, since after it is processed once, the corresponding worker will free it and make a mess.