bugfix: Transparently hashing an empty input is not possible in FULL_RW mode.
This is possible in APPEND_ONLY mode because we use indexes in the blocks and therefore we can start with 1. When running in FULL_RW, we use nonces and we don’t get one if no block was inserted, so we can’t just generate a random nonce because then the algorithm would be undeterministic for empty inputs. Therefore, ishake_final() must only consider performing one last hash if we are running in APPEND_ONLY mode.

If we want to hash an empty input in FULL_RW mode, the user is responsible for calling ishake_insert() with a block that contains a nonce and no data. Then the user controls the nonce and the algorithm can deterministically hash an empty input.